Split MergeCanvas into physics, renderer, and config modules
MergeCanvas.tsx mixed the collision/gravity physics engine, canvas drawing
code, and tuning constants in one 280-line component. Extracted
numberMergePhysics.ts (circlePair/resolveCollision/stepPhysics),
numberMergeRenderer.ts (drawBall/drawWalls/renderFrame), and
numberMergeCanvasConfig.ts (dimensions, physics constants, color/Hebrew
number maps), alongside the existing numberMergeStore.ts. MergeCanvas.tsx
is now 89 lines of state sync, tick-loop composition, and pointer
handlers. No behavior change.
