+++ gamesformykids/app/games/number-merge/components/MergeCanvas.tsx
@@ -1,170 +1,11 @@
 'use client';
 import { useRef, useEffect, useCallback } from 'react';
-import { useNumberMergeStore, Ball, RADIUS_FOR_VALUE } from '../numberMergeStore';
+import { useNumberMergeStore, type Ball } from '../numberMergeStore';
 import { speakHebrew } from '@/lib/utils/speech/speaker';
 import { useCanvasLoop } from '@/hooks/canvas/useCanvasLoop';
-
-const CANVAS_W = 320;
-const CANVAS_H = 480;
-const FLOOR_Y = CANVAS_H - 20;
-const WALL_LEFT = 10;
-const WALL_RIGHT = CANVAS_W - 10;
-const GRAVITY = 0.4;
-const DAMPING = 0.55;
-const FRICTION = 0.85;
-
-const NUMBER_COLORS: Record<number, string> = {
-  1:  '#FF6B6B',
-  2:  '#FF8E53',
-  3:  '#FFCA28',
-  4:  '#66BB6A',
-  5:  '#26C6DA',
-  6:  '#42A5F5',
-  7:  '#7E57C2',
-  8:  '#EC407A',
-  9:  '#8D6E63',
-  10: '#455A64',
-};
-
-const HEBREW_NUMBERS: Record<number, string> = {
-  1:'אחד',2:'שניים',3:'שלוש',4:'ארבע',5:'חמש',
-  6:'שש',7:'שבע',8:'שמונה',9:'תשע',10:'עשר',
-};
-
-function circlePair(a: Ball, b: Ball) {
-  const dx = b.x - a.x;
-  const dy = b.y - a.y;
-  const dist = Math.sqrt(dx * dx + dy * dy);
-  const minDist = a.radius + b.radius;
-  return { dx, dy, dist, minDist, overlap: minDist - dist };
-}
-
-function resolveCollision(balls: Ball[]): { balls: Ball[]; merges: { x: number; y: number; value: number }[] } {
-  const merges: { x: number; y: number; value: number }[] = [];
-  const toRemove = new Set<number>();
-
-  for (let i = 0; i < balls.length; i++) {
-    for (let j = i + 1; j < balls.length; j++) {
-      const a = balls[i];
-      const b = balls[j];
-      if (!a || !b || toRemove.has(a.id) || toRemove.has(b.id)) continue;
-      if (a.value !== b.value) continue;
-
-      const { dist, minDist } = circlePair(a, b);
-      if (dist >= minDist) continue;
-
-      // Same value touching → merge
-      toRemove.add(a.id);
-      toRemove.add(b.id);
-      const newVal = Math.min(a.value + b.value, 10);
-      const nx = (a.x + b.x) / 2;
-      const ny = (a.y + b.y) / 2;
-      const merged: Ball = {
-        id: Date.now() + Math.random(),
-        x: nx,
-        y: ny,
-        vx: (a.vx + b.vx) * 0.5,
-        vy: (a.vy + b.vy) * 0.5,
-        value: newVal,
-        radius: RADIUS_FOR_VALUE(newVal),
-        merging: false,
-        merged: false,
-      };
-      balls.push(merged);
-      merges.push({ x: nx, y: ny, value: newVal });
-      // skip inner j loop since a is gone
-      break;
-    }
-  }
-
-  return {
-    balls: balls.filter(b => !toRemove.has(b.id)),
-    merges,
-  };
-}
-
-function stepPhysics(balls: Ball[]): { balls: Ball[]; merges: { x: number; y: number; value: number }[] } {
-  // Apply gravity + velocity
-  for (const b of balls) {
-    b.vy += GRAVITY;
-    b.x += b.vx;
-    b.y += b.vy;
-
-    // Floor
-    if (b.y + b.radius >= FLOOR_Y) {
-      b.y = FLOOR_Y - b.radius;
-      b.vy *= -DAMPING;
-      b.vx *= FRICTION;
-    }
-    // Walls
-    if (b.x - b.radius <= WALL_LEFT) {
-      b.x = WALL_LEFT + b.radius;
-      b.vx *= -DAMPING;
-    }
-    if (b.x + b.radius >= WALL_RIGHT) {
-      b.x = WALL_RIGHT - b.radius;
-      b.vx *= -DAMPING;
-    }
-  }
-
-  // Ball-ball separation (push apart first, then merge)
-  for (let i = 0; i < balls.length; i++) {
-    for (let j = i + 1; j < balls.length; j++) {
-      const a = balls[i];
-      const b = balls[j];
-      if (!a || !b) continue;
-      const { dist, dx, dy, overlap } = circlePair(a, b);
-      if (overlap <= 0 || dist === 0) continue;
-      if (a.value === b.value) continue; // will be merged separately
-
-      const nx = dx / dist;
-      const ny = dy / dist;
-      const push = overlap / 2;
-      a.x -= nx * push;
-      a.y -= ny * push;
-      b.x += nx * push;
-      b.y += ny * push;
-
-      // Elastic-ish bounce
-      const dvx = b.vx - a.vx;
-      const dvy = b.vy - a.vy;
-      const dot = dvx * nx + dvy * ny;
-      if (dot < 0) {
-        const impulse = dot * 0.5;
-        a.vx += impulse * nx;
-        a.vy += impulse * ny;
-        b.vx -= impulse * nx;
-        b.vy -= impulse * ny;
-      }
-    }
-  }
-
-  return resolveCollision(balls);
-}
-
-function drawBall(ctx: CanvasRenderingContext2D, ball: Ball) {
-  const color = NUMBER_COLORS[ball.value] ?? '#999';
-  ctx.beginPath();
-  ctx.arc(ball.x, ball.y, ball.radius, 0, Math.PI * 2);
-  ctx.fillStyle = color;
-  ctx.fill();
-  ctx.strokeStyle = 'rgba(0,0,0,0.2)';
-  ctx.lineWidth = 2;
-  ctx.stroke();
-
-  ctx.fillStyle = '#fff';
-  ctx.font = `bold ${Math.max(10, ball.radius * 0.85)}px Arial`;
-  ctx.textAlign = 'center';
-  ctx.textBaseline = 'middle';
-  ctx.fillText(String(ball.value), ball.x, ball.y);
-}
-
-function drawWalls(ctx: CanvasRenderingContext2D) {
-  ctx.fillStyle = '#d1a96a';
-  ctx.fillRect(0, FLOOR_Y, CANVAS_W, CANVAS_H - FLOOR_Y);
-  ctx.fillRect(0, 0, WALL_LEFT, CANVAS_H);
-  ctx.fillRect(WALL_RIGHT, 0, CANVAS_W - WALL_RIGHT, CANVAS_H);
-}
+import { CANVAS_W, CANVAS_H, WALL_LEFT, WALL_RIGHT, HEBREW_NUMBERS } from '../numberMergeCanvasConfig';
+import { stepPhysics } from '../numberMergePhysics';
+import { renderFrame } from '../numberMergeRenderer';
 
 export default function MergeCanvas() {
   const ballsRef = useRef<Ball[]>([]);
@@ -212,39 +53,7 @@ export default function MergeCanvas() {
       setBalls(newBalls.map(b => ({ ...b })));
     }
 
-    // Draw
-    ctx.clearRect(0, 0, CANVAS_W, CANVAS_H);
-    // Background
-    ctx.fillStyle = '#FFF8E1';
-    ctx.fillRect(0, 0, CANVAS_W, CANVAS_H);
-    drawWalls(ctx);
-
-    // Drop indicator line
-    ctx.setLineDash([4, 4]);
-    ctx.strokeStyle = 'rgba(0,0,0,0.15)';
-    ctx.lineWidth = 1;
-    ctx.beginPath();
-    ctx.moveTo(dropX, 0);
-    ctx.lineTo(dropX, FLOOR_Y);
-    ctx.stroke();
-    ctx.setLineDash([]);
-
-    // Preview ball at top
-    const previewR = RADIUS_FOR_VALUE(nextValue);
-    const previewColor = NUMBER_COLORS[nextValue] ?? '#999';
-    ctx.globalAlpha = 0.5;
-    ctx.beginPath();
-    ctx.arc(dropX, previewR + 2, previewR, 0, Math.PI * 2);
-    ctx.fillStyle = previewColor;
-    ctx.fill();
-    ctx.fillStyle = '#fff';
-    ctx.font = `bold ${Math.max(10, previewR * 0.85)}px Arial`;
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
-    ctx.fillText(String(nextValue), dropX, previewR + 2);
-    ctx.globalAlpha = 1;
-
-    for (const b of newBalls) drawBall(ctx, b);
+    renderFrame(ctx, { balls: newBalls, dropX, nextValue });
   }, [dropX, nextValue, addScore, setBalls, triggerMergeFlash, endGame, announceEquation]);
 
   const canvasRef = useCanvasLoop(tick);
