+++ gamesformykids/app/games/number-merge/numberMergeCanvasConfig.ts
@@ -1,0 +1,26 @@
+export const CANVAS_W = 320;
+export const CANVAS_H = 480;
+export const FLOOR_Y = CANVAS_H - 20;
+export const WALL_LEFT = 10;
+export const WALL_RIGHT = CANVAS_W - 10;
+export const GRAVITY = 0.4;
+export const DAMPING = 0.55;
+export const FRICTION = 0.85;
+
+export const NUMBER_COLORS: Record<number, string> = {
+  1:  '#FF6B6B',
+  2:  '#FF8E53',
+  3:  '#FFCA28',
+  4:  '#66BB6A',
+  5:  '#26C6DA',
+  6:  '#42A5F5',
+  7:  '#7E57C2',
+  8:  '#EC407A',
+  9:  '#8D6E63',
+  10: '#455A64',
+};
+
+export const HEBREW_NUMBERS: Record<number, string> = {
+  1:'אחד',2:'שניים',3:'שלוש',4:'ארבע',5:'חמש',
+  6:'שש',7:'שבע',8:'שמונה',9:'תשע',10:'עשר',
+};
